requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-autocomplete-light==3.9.7
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-dajax==0.9.2
django-dajaxice==0.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-model-utils==4.3.1
django-timezone-field==6.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
eventlet==0.33.3
gevent==23.9.1
greenlet==3.0.0rc3
idna==3.4
kombu==5.3.2
oauthlib==3.2.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
pygame==2.5.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
